+++ src/util.js
@@ -1,12 +1,5 @@
 import {RenderPosition} from './const.js';
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
 const humanizeDate = (date) => {
   const month = date.toLocaleString(`en-US`, {month: `short`});
   const day = date.getDate();
@@ -32,16 +25,8 @@ const createElement = (template) => {
   return newElement.firstChild;
 };
 
-const getRandomArrayItem = (array) =>
-  array[Math.floor(Math.random() * array.length)];
-
-const getRandomBoolean = () => Math.random() > 0.5;
-
 export {
   render,
   createElement,
-  getRandomArrayItem,
-  getRandomInteger,
-  getRandomBoolean,
   humanizeDate
 };
